Restructured Files
Reorganized the File Tree for clairty
