+++ src/gui/FunctionPanel.java
@@ -1,4 +1,4 @@
-package gui;
+package src.gui;
 
 import java.util.HashMap;
 import javax.swing.JButton;
