+++ src/gui/OperationsPanel.java
@@ -1,4 +1,4 @@
-package gui;
+package src.gui;
 
 import java.util.*;
 import javax.swing.*;
