+++ src/utils/CalcApp.java
@@ -1,4 +1,4 @@
-package src;
+package src.utils;
 
 public class CalcApp {
     //-------------------- Vars
