+++ src/utils/CalcController.java
@@ -1,4 +1,4 @@
-package src;
+package src.utils;
 import java.math.*;
 import java.text.DecimalFormat;
