+++ src/utils/MyMath.java
@@ -1,4 +1,4 @@
-package src;
+package src.utils;
 import java.util.*;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
