+++ unitTests/ModelTest.java
@@ -1,9 +1,10 @@
 
 package unitTests;
 
-import src.CalcModel;
 import java.math.BigDecimal;
 import java.util.HashMap;
+
+import src.utils.CalcModel;
 
 public class ModelTest {
